Toasts, Dialog, Autocomplete e Combobox

+++ webiv-front-angular/src/app/view/departamento/departamento-search/departamento-search.component.ts
@@ -4,6 +4,7 @@ import { Router, ActivatedRoute } from '@angular/router';
 import { DepartamentoService } from 'src/app/service/departamento.service';
 import { TipoAcaoValues } from 'src/app/model/tipo-acao';
 import { MessagesService } from 'src/app/service/messages.service';
+import { TdDialogService } from '@covalent/core/dialogs';
 
 @Component({
   selector: 'app-departamento-search',
@@ -26,7 +27,9 @@ export class DepartamentoSearchComponent implements OnInit {
   constructor(private router: Router,
             private activatedRoute: ActivatedRoute,
             private departamentoService: DepartamentoService,
-            private messageService: MessagesService) { 
+            private messageService: MessagesService,
+            private _dialogService: TdDialogService,
+            private _viewContainerRef: ViewContainerRef) { 
   }
 
   /**
@@ -77,15 +80,36 @@ export class DepartamentoSearchComponent implements OnInit {
     });
   }
     
+  /**
+   * Método para remover um departamento
+   */
   remover(id: number){
-    this.departamentoService.remover(id).subscribe(dados => {
-      this.messageService.toastSuccess('Departamento excluído com sucesso.');
-      this.listar();
-    },
-    (error: any) => {
-      console.log(error.error.message);
-      this.messageService.toastError(error.error.message);
-      
+    this.openRemoverConfirm(id);
+  }
+
+  openRemoverConfirm(id: number): void {
+    this._dialogService.openConfirm({
+      message: 'Tem certeza que deseja excluir esse departamento?',
+      disableClose: true, // defaults to false
+      viewContainerRef: this._viewContainerRef, //OPTIONAL
+      title: 'Excluir departamento', //OPTIONAL, hides if not provided
+      cancelButton: 'Não', //OPTIONAL, defaults to 'CANCEL'
+      acceptButton: 'Sim', //OPTIONAL, defaults to 'ACCEPT'
+      width: '500px', //OPTIONAL, defaults to 400px
+    }).afterClosed().subscribe((accept: boolean) => {
+      if (accept) {
+        this.departamentoService.remover(id).subscribe(dados => {
+          this.messageService.toastSuccess('Departamento excluído com sucesso.');
+          this.listar();
+        },
+        (error: any) => {
+          console.log(error.error.message);
+          this.messageService.toastError(error.error.message);
+          
+        });
+      } else {
+        // DO SOMETHING ELSE
+      }
     });
   }
 }
